requirements.txt
torch
requests
pillow
